reorganized backend + added queryVehicles function to text chatbot
