+++ .gitignore
@@ -219,8 +219,6 @@ __azurite_db*__.json
 .python_packages
 
 
-.funcignore
-
 #React files
 
 # Logs
